+++ conjuror/plans/beam.py
@@ -455,8 +455,10 @@ class Beam(Generic[TMachine], BeamDynamicsMixin, BeamVisualizationMixin, ABC):
             The couch rotation.
         """
 
-        if len(beam_name) > 16:
-            raise ValueError("Beam name must be less than or equal to 16 characters")
+        if len(beam_name) > 64:
+            # BeamName's can accept up to 64 characters:
+            # https://dicom.innolitics.com/ciods/rt-plan/rt-beams/300a00b0/300a00c2
+            raise ValueError("Beam name must be less than or equal to 64 characters")
 
         # Private attributes used for dicom creation only
         self._fluence_mode = fluence_mode